# back-end/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorLogResponse:
    description: Standard error response body containing details about the error occurrence
    properties:
      call:
        description: The API endpoint/URL that was called
        example: /getpoint/123
        type: string
      debuginfo:
        description: Detailed debug information about the error
        example: Point with ID 123 not found
        type: string
      error:
        description: Short error title/description
        example: Not Found
        type: string
      originator:
        description: IP address or identifier of the request originator
        example: 192.168.1.100
        type: string
      return_code:
        description: HTTP status code returned
        example: 404
        type: integer
      timeref:
        description: 'Timestamp when the error occurred (format: YYYY-MM-DD HH:MM)'
        example: 2025-12-25 14:30
        type: string
    type: object
...
  handlers.PointDetailDTO:
    description: Detailed response containing all information about a specific charging
      point
    properties:
      cap:
        description: Maximum charging capacity in kW
        example: 22
        type: integer
      is_manual_price:
        description: Whether the price is manually set or dynamically calculated
        example: false
        type: boolean
      kwhprice:
        description: Current price per kWh in local currency
        example: 0.35
        type: number
      lat:
        description: Latitude coordinate of the charging point location
        example: "37.983810"
        type: string
      lon:
        description: Longitude coordinate of the charging point location
        example: "23.727539"
        type: string
      pointid:
        description: Unique identifier of the charging point
        example: "123"
        type: string
      reservationendtime:
        description: End time of current reservation (current time if not reserved)
        example: 2025-12-25 14:30
        type: string
      status:
        description: Current status of the charging point (available, charging, reserved,
          malfunction, offline)
        example: available
        type: string
    type: object
...
  handlers.ReserveResponseDTO:
    description: Response containing the result of a reservation request
    properties:
      pointid:
        description: ID of the charging point
        example: "123"
        type: string
      reservationendtime:
        description: 'End time of the reservation (format: YYYY-MM-DD HH:MM), or ''1970-01-01
          00:00'' if reservation failed'
        example: 2025-12-25 15:00
        type: string
      status:
        description: |-
          Result status: 'reserved' if successful, 'not_found' if point doesn't exist,
          or current status if point is not available for reservation
        example: reserved
        type: string
    type: object
...
  handlers.StatusChangeDTO:
    description: Record of a status change event including timestamp and state transition
    properties:
      new_state:
        description: New status after the change
        example: charging
        type: string
      old_state:
        description: Previous status before the change
        example: available
        type: string
      timeref:
        description: 'Timestamp when the status change occurred (format: YYYY-MM-DD
          HH:MM)'
        example: 2025-12-25 14:30
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing electric vehicle charging points
  title: EV Charging API
  version: "1.0"
paths:
...
  /getpoint/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves detailed information about a specific charging point including
        location coordinates, current status, capacity, pricing, and reservation details.
      parameters:
      - description: Charging Point ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved charging point details
          schema:
            $ref: '#/definitions/handlers.PointDetailDTO'
        "404":
          description: Not Found - Charging point does not exist
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "500":
          description: Internal Server Error - Database error
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
      summary: Get charging point by ID
      tags:
      - Points
...
  /pointstatus/{pointid}/{from}/{to}:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves the history of status changes for a specific charging point within a date range.
        Status changes are logged whenever a point transitions between states (e.g., available → charging).
        Results can be returned in JSON or CSV format.
      parameters:
      - description: Charging Point ID
        in: path
        name: pointid
        required: true
        type: string
      - description: Start date in YYYYMMDD format
        example: "20251201"
        in: path
        name: from
        required: true
        type: string
      - description: End date in YYYYMMDD format (inclusive)
        example: "20251225"
        in: path
        name: to
        required: true
        type: string
      - default: json
        description: Response format (json or csv)
        in: query
        name: format
        type: string
      produces:
      - application/json
      - text/csv
      responses:
        "200":
          description: Successfully retrieved status change history
          schema:
            items:
              $ref: '#/definitions/handlers.StatusChangeDTO'
            type: array
        "204":
          description: No Content - No status changes found in the specified date
            range
        "400":
          description: Bad Request - Invalid date format
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "404":
          description: Not Found - Charging point does not exist
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "500":
          description: Internal Server Error - Database error
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
      summary: Get status change history for a charging point
      tags:
      - Points
  /reserve/{id}:
    post:
      consumes:
      - application/json
      description: |-
        Attempts to reserve a charging point for a specified duration.
        The reservation duration must be between 15 and 60 minutes (defaults to 30 if not specified).
        Only 'available' charging points can be reserved.
        The operation is atomic and uses row-level locking to prevent race conditions.
      parameters:
      - description: Charging Point ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reservation result (check status field for success/failure
            reason)
          schema:
            $ref: '#/definitions/handlers.ReserveResponseDTO'
        "400":
          description: Bad Request - Invalid point ID
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "500":
          description: Internal Server Error - Database error
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
      summary: Reserve a charging point
      tags:
      - Reservations
  /reserve/{id}/{minutes}:
    post:
      consumes:
      - application/json
      description: |-
        Attempts to reserve a charging point for a specified duration.
        The reservation duration must be between 15 and 60 minutes (defaults to 30 if not specified).
        Only 'available' charging points can be reserved.
        The operation is atomic and uses row-level locking to prevent race conditions.
      parameters:
      - description: Charging Point ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Reservation duration in minutes (15-60, default: 30)'
        in: path
        name: minutes
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation result (check status field for success/failure
            reason)
          schema:
            $ref: '#/definitions/handlers.ReserveResponseDTO'
        "400":
          description: Bad Request - Invalid point ID
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
        "500":
          description: Internal Server Error - Database error
          schema:
            $ref: '#/definitions/handlers.ErrorLogResponse'
      summary: Reserve a charging point
      tags:
      - Reservations
...
swagger: "2.0"